--- src/components/common/input/AuthInput.tsx
@@ -1,18 +1,0 @@
-import type { LoginSchemeType } from '../../../models/auth';
-import * as S from './AuthInput.styled';
-import { type UseFormRegister } from 'react-hook-form';
-
-interface AuthInputProps {
-  name: 'email' | 'password';
-  register: UseFormRegister<LoginSchemeType>;
-  placeholder: string;
-  type: 'text' | 'password';
-}
-
-const AuthInput = ({ name, register, placeholder, type }: AuthInputProps) => {
-  return (
-    <S.Container type={type} placeholder={placeholder} {...register(name)} />
-  );
-};
-
-export default AuthInput;
